Refactor abstract base classes to use more detailed schema
The AbstractDataExtensionObject now defines a schema as a nested
associative array that specifies the following for each field:
* is_primary ... If the field is a primary key this is true
* is_required .. If the field is not nullable this is true
* type ......... This is the field type expressed as a constant that is
                 defined in the class.

Depending on the type and the other properties it is now possible to set
a field to null if it previously had a value.
At the moment only fields that are not required, not primary and of type
TEXT can be set to null. Any other field will be ignored by the save
operation and will retain its original value.
To set a field to null, set its value to null or ''.

Closes #1